Made dashboard responsive and added cards
