+++ src/components/pages/Dashboard/ExpenseBarChart.js
@@ -13,14 +13,14 @@ import {
 
 const ExpenseBarChart = ({ groupedExpenses, CATEGORY_COLORS }) => {
   return (
-    <ResponsiveContainer width='100%' height={300}>
+    <ResponsiveContainer width='100%' height={200}>
       <BarChart
         data={groupedExpenses}
         margin={{
           top: 5,
           right: 30,
           left: 20,
-          bottom: 50,
+          bottom: 5,
         }}
       >
         <CartesianGrid vertical={false} fill='gray' fillOpacity={0.1} />
